项目管理设定模块

+++ exchange/src/main/java/cn/edu/cmu/framework/util/MaxNumUtils.java
@@ -14,10 +14,10 @@ public class MaxNumUtils {
 
 
 
-    public static  String getMaxNum(String key,Long year, int limit) throws Exception {
+    public static  String getMaxNum(String key,int year, int limit) throws Exception {
 
         MaxCodeService service = WebAppContextUtils.getBean(MaxCodeService.class);
-        long max = service.getMaxCode(key, year);
+        long max = service.getMaxCode(key, (long)year);
 
 
         if(limit-String.valueOf(max).length()<0){
